generators/EMADL2CPP/extractor/requirements.txt
python-gitlab
GitPython